chore: rename files to match other file's naming convention
